refactor: -Revue des méthodes de traitements d'exception et d'IntegrityError dans RepositoriesUtils -Deplacement de workers_task_names.py dans le module des tasks
